moved builtin_segments under builtins mod
